Split the application into multiple files Fix #4
